Restrict system deletion

+++ components/director/internal/domain/application/service.go
@@ -3,6 +3,7 @@ package application
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
 
@@ -30,7 +31,7 @@ const (
 
 type repoCreatorFunc func(ctx context.Context, application *model.Application) error
 
-//go:generate mockery -name=ApplicationRepository -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=ApplicationRepository --output=automock --outpkg=automock --case=underscore
 type ApplicationRepository interface {
 	Exists(ctx context.Context, tenant, id string) (bool, error)
 	GetByID(ctx context.Context, tenant, id string) (*model.Application, error)
@@ -45,7 +46,7 @@ type ApplicationRepository interface {
 	DeleteGlobal(ctx context.Context, id string) error
 }
 
-//go:generate mockery -name=LabelRepository -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=LabelRepository --output=automock --outpkg=automock --case=underscore
 type LabelRepository interface {
 	GetByKey(ctx context.Context, tenant string, objectType model.LabelableObject, objectID, key string) (*model.Label, error)
 	ListForObject(ctx context.Context, tenant string, objectType model.LabelableObject, objectID string) (map[string]*model.Label, error)
@@ -53,39 +54,40 @@ type LabelRepository interface {
 	DeleteAll(ctx context.Context, tenant string, objectType model.LabelableObject, objectID string) error
 }
 
-//go:generate mockery -name=WebhookRepository -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=WebhookRepository --output=automock --outpkg=automock --case=underscore
 type WebhookRepository interface {
 	CreateMany(ctx context.Context, items []*model.Webhook) error
 }
 
-//go:generate mockery -name=RuntimeRepository -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=RuntimeRepository --output=automock --outpkg=automock --case=underscore
 type RuntimeRepository interface {
 	Exists(ctx context.Context, tenant, id string) (bool, error)
+	ListAll(ctx context.Context, tenantID string, filter []*labelfilter.LabelFilter) ([]*model.Runtime, error)
 }
 
-//go:generate mockery -name=IntegrationSystemRepository -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=IntegrationSystemRepository --output=automock --outpkg=automock --case=underscore
 type IntegrationSystemRepository interface {
 	Exists(ctx context.Context, id string) (bool, error)
 }
 
-//go:generate mockery -name=LabelUpsertService -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=LabelUpsertService --output=automock --outpkg=automock --case=underscore
 type LabelUpsertService interface {
 	UpsertMultipleLabels(ctx context.Context, tenant string, objectType model.LabelableObject, objectID string, labels map[string]interface{}) error
 	UpsertLabel(ctx context.Context, tenant string, labelInput *model.LabelInput) error
 }
 
-//go:generate mockery -name=ScenariosService -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=ScenariosService --output=automock --outpkg=automock --case=underscore
 type ScenariosService interface {
 	EnsureScenariosLabelDefinitionExists(ctx context.Context, tenant string) error
 	AddDefaultScenarioIfEnabled(ctx context.Context, labels *map[string]interface{})
 }
 
-//go:generate mockery -name=UIDService -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=UIDService --output=automock --outpkg=automock --case=underscore
 type UIDService interface {
 	Generate() string
 }
 
-//go:generate mockery -name=ApplicationHideCfgProvider -output=automock -outpkg=automock -case=underscore
+//go:generate mockery --name=ApplicationHideCfgProvider --output=automock --outpkg=automock --case=underscore
 type ApplicationHideCfgProvider interface {
 	GetApplicationHideSelectors() (map[string][]string, error)
 }
@@ -301,6 +303,21 @@ func (s *service) Delete(ctx context.Context, id string) error {
 		return errors.Wrapf(err, "while loading tenant from context")
 	}
 
+	formations, runtimes, err := s.isInFormationAndRuntime(ctx, appTenant, id)
+	if err != nil {
+		return errors.Wrapf(err, "while blah blah")
+	}
+
+	if len(formations) > 0 {
+		msg := fmt.Sprintf("System deletion failed: the system is part of a formation - %s", strings.Join(formations, ", "))
+
+		if len(runtimes) > 0 {
+			msg = fmt.Sprintf("%s and of runtime - %s", msg, strings.Join(runtimes, ", "))
+		}
+
+		return apperrors.NewInvalidOperationError(msg)
+	}
+
 	err = s.appRepo.Delete(ctx, appTenant, id)
 	if err != nil {
 		return errors.Wrapf(err, "while deleting Application with id %s", id)
@@ -513,4 +530,52 @@ func (s *service) ensureIntSysExists(ctx context.Context, id *string) (bool, err
 	}
 	log.C(ctx).Infof("Integration System with id %s exists", *id)
 	return true, nil
+}
+
+func (s *service) isInFormationAndRuntime(ctx context.Context, tenant, applicationID string) ([]string, []string, error) {
+	log.C(ctx).Infof("Getting formations for application with id %s", applicationID)
+
+	applicationLabel, err := s.GetLabel(ctx, applicationID, model.ScenariosKey)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	scenarios, err := label.ValueToStringsSlice(applicationLabel.Value)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "while parsing application label values")
+	}
+
+	if len(scenarios) == 0 {
+		return nil, nil, nil
+	}
+
+	scenariosQuery := buildQueryForScenarios(scenarios)
+	runtimeScenariosFilter := []*labelfilter.LabelFilter{labelfilter.NewForKeyWithQuery(model.ScenariosKey, scenariosQuery)}
+
+	log.C(ctx).Infof("Listing runtimes matching the query %s", scenariosQuery)
+	runtimes, err := s.runtimeRepo.ListAll(ctx, tenant, runtimeScenariosFilter)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "while getting runtimes")
+	}
+
+	var runtimesNames []string
+	for _, r := range runtimes {
+		runtimesNames = append(runtimesNames, r.Name)
+	}
+
+	return scenarios, runtimesNames, nil
+}
+
+func buildQueryForScenarios(scenarios []string) string {
+	var queryBuilder strings.Builder
+	for idx, scenario := range scenarios {
+		if idx > 0 {
+			queryBuilder.WriteString(` || `)
+		}
+
+		queryBuilder.WriteString(fmt.Sprintf(`@ == "%s"`, scenario))
+	}
+	query := fmt.Sprintf(`$[*] ? ( %s )`, queryBuilder.String())
+
+	return query
 }
